refactor history to deal with less code
